🚚 SPEAKERS: Move <Product /> to components folder
